+++ src/components/autocomplete/index.ts
@@ -1,0 +1,6 @@
+export * from "./Autocomplete"
+export type {
+  AutocompleteProps,
+  AutocompleteItem,
+  AutocompleteChangeEvent,
+} from "./Autocomplete.model"