+++ src/components/tabs/Tabs.lazy.tsx
@@ -1,0 +1,13 @@
+import { createLazyWrapper } from "@/helpers/createLazyWrapper"
+
+import TabsClient from "./Tabs.client"
+import TabsServer from "./Tabs.server"
+
+import type { TabsProps } from "./Tabs.model"
+
+export default createLazyWrapper<TabsProps>({
+  name: "Tabs",
+  Client: TabsClient,
+  Server: TabsServer,
+  isInteractive: () => true,
+})